Add Errors and rename CustomError & fixed inheritance
CustomError => ApiError
added ForbiddenError
Added UnauthorizedError
Fixed inheritance but
