+++ src/errors/client/NotFoundError.ts
@@ -1,5 +1,5 @@
+import { type ErrorOptions } from "../ApiError";
 import { BadRequestError } from "./BadRequestError";
-import { type ErrorOptions } from "./CustomError";
 
 export class NotFoundError extends BadRequestError {
   protected readonly _statusCode = 404;
@@ -8,6 +8,6 @@ export class NotFoundError extends BadRequestError {
     super(message, options);
 
     // Only because we are extending a built in class
-    Object.setPrototypeOf(this, NotFoundError?.prototype);
+    Object.setPrototypeOf(this, NotFoundError.prototype);
   }
 }
