+++ src/errors/index.ts
@@ -1,4 +1,15 @@
 // barrel exports so that all errors can be imported from a single path
-export * from "./BadRequestError";
-export * from "./NotFoundError";
-export * from "./ServerError";
+
+// base
+export * from "./ApiError";
+
+// 400
+export * from "./client/BadRequestError";
+export * from "./client/ForbiddenError";
+export * from "./client/NotFoundError";
+export * from "./client/UnauthorizedError";
+
+export * from "./client/ValidationError";
+
+// 500
+export * from "./server/ServerError";
